speech-service/requirements.txt
fastapi
uvicorn[standard]
websockets
torch
numpy
openai-whisper
soundfile
ffmpeg-python
certifi
pyopenssl
